requirements.txt
beautifulsoup4==4.11.1
fake-useragent==1.1.1
lxml==4.9.1
python-decouple==3.6
requests==2.28.1